requirements.txt
sentence-transformers==2.7.0
psycopg2-binary==2.9.9
PyMuPDF==1.24.0
numpy==1.26.4
scikit-learn==1.4.2
python-dotenv==1.0.1
